Insert works and has been tested, also spread the testing monolith into multiple files
